+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 # Change Log
 
+## 0.1.6 (July 30, 2025)
+
+Notes: 
+* tbr-deal-finder now also tracks deals on the books in your wishlist. Works for all retailers.   
+
+BUG FIXES:
+* Fixed issue where no deals would display if libro is the only tracked audiobook retailer.
+* Fixed retailer cli setup forcing a user to select at least two audiobook retailers.
+
 ## 0.1.5 (July 30, 2025)
 
 Notes: